+++ src/Resources/app/administration/src/module/heptacom-admin-open-auth-client/snippets/index.js
@@ -1,0 +1,7 @@
+import deDE from './de-DE';
+import enGB from './en-GB';
+
+export default {
+    'de-DE': deDE,
+    'en-GB': enGB,
+};