+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
+* Add configuration in administration
 * Add custom login grant